Rename namespace ForEvolve.Azure.Storage.Object to ForEvolve.Azure.Storage.Blob
