+++ src/ForEvolve.Azure/Storage/Blob/IFileRepository.cs
@@ -8,7 +8,7 @@ using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ForEvolve.Azure.Storage.Object
+namespace ForEvolve.Azure.Storage.Blob
 {
     public interface IFileRepository
     {
