+++ src/ForEvolve.Azure/Storage/Blob/IObjectStorageRepository.cs
@@ -6,7 +6,7 @@ using Newtonsoft.Json;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ForEvolve.Azure.Storage.Object
+namespace ForEvolve.Azure.Storage.Blob
 {
     public interface IObjectStorageRepository<TModel>
     {
