+++ src/ForEvolve.Azure/Storage/Blob/IObjectStorageSettings.cs
@@ -1,4 +1,4 @@
-﻿namespace ForEvolve.Azure.Storage.Object
+﻿namespace ForEvolve.Azure.Storage.Blob
 {
     public interface IObjectStorageSettings : IStorageSettings
     {
