+++ src/ForEvolve.Azure/Storage/Blob/ObjectStorageSettings.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace ForEvolve.Azure.Storage.Object
+namespace ForEvolve.Azure.Storage.Blob
 {
     public class ObjectStorageSettings : StorageSettings, IObjectStorageSettings
     {
